+++ examples/storybook/.storybook/webpack.config.js
@@ -1,18 +1,24 @@
 const wixStorybookConfig = require('yoshi/config/webpack.config.storybook');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
 
 module.exports = ({ config }) => {
   const baseConfig = wixStorybookConfig(config);
   return {
     ...baseConfig,
-    plugins: [
-      ...baseConfig.plugins,
-      new CopyWebpackPlugin([
+    module: {
+      ...baseConfig.module,
+      rules: [
         {
-          from: 'node_modules/wix-rich-content-plugin-html/dist/statics/',
-          to: 'static/',
+          test: /\.m?js$/,
+          exclude: /(node_modules|bower_components)/,
+          use: {
+            loader: 'babel-loader',
+            options: {
+              plugins: ['babel-plugin-preval'],
+            },
+          },
         },
-      ]),
-    ],
+        ...baseConfig.module.rules,
+      ],
+    },
   };
 };