+++ examples/storybook/stories/Components/SourceCode.js
@@ -1,5 +1,13 @@
+/* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable jsx-a11y/click-events-have-key-events */
 import React from 'react';
 import PropTypes from 'prop-types';
+import cx from 'classnames';
+
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+
+import { atomDark } from 'react-syntax-highlighter/dist/esm/styles/prism';
+
 import styles from './SourceCode.scss';
 
 class SourceCode extends React.Component {
@@ -14,11 +22,25 @@ class SourceCode extends React.Component {
     const { code } = this.props;
     const { open } = this.state;
 
-    if (open) {
-      return <div className={styles.sourcecode}>{code}</div>;
-    } else {
-      return <div>See Source Code</div>;
-    }
+    return (
+      <div>
+        <div className={cx(styles.sourcecode, { [styles.open]: open })}>
+          <SyntaxHighlighter language="javascript" style={atomDark}>
+            {code}
+          </SyntaxHighlighter>
+        </div>
+
+        <div
+          onClick={() =>
+            this.setState(_state => ({
+              open: !_state.open,
+            }))
+          }
+        >
+          {'</>'} {open ? 'Hide' : 'show'} Code
+        </div>
+      </div>
+    );
   }
 }