+++ examples/storybook/stories/Components/ViewerWrapper.js
@@ -50,6 +50,7 @@ const configs = {
 };
 
 const plugins = [
+  //dummy
   pluginButton(),
   pluginCodeBlock(),
   pluginDivider(),
@@ -79,7 +80,7 @@ const ViewerWrapper = ({ contentState, palette }) => {
 
 ViewerWrapper.propTypes = {
   contentState: PropTypes.object,
-  palette: PropTypes.arrayOf(PropTypes.string),
+  palette: PropTypes.arrayOf(PropTypes.object),
 };
 
 export default ViewerWrapper;