+++ packages/plugin-hashtag/web/src/viewer.js
@@ -1,20 +1,16 @@
 import { hasLinksInBlock } from 'wix-rich-content-common';
 import { default as createHashtagDecorator } from './HashtagDecorator';
+import { HASHTAG_TYPE } from './types';
+import { DEFAULTS, THEME as theme } from './defaults';
 const HashtagDecorator = createHashtagDecorator(hasLinksInBlock);
 export { HashtagDecorator };
 
-import { HASHTAG_TYPE } from './types';
-import { DEFAULTS } from './defaults';
-
 export const pluginHashtag = (config = {}) => {
+  const pluginConfig = { ...DEFAULTS.config, ...config };
   return {
-    config: { ...DEFAULTS.config, ...config },
+    config: pluginConfig,
     type: HASHTAG_TYPE,
-    decorator: HashtagDecorator,
-    theme: colors => ({
-      hashtag: {
-        color: colors.actionColor,
-      },
-    }),
+    decorator: theme => new HashtagDecorator({ ...pluginConfig, ...theme }),
+    theme,
   };
 };