+++ packages/plugin-link/web/src/editor.js
@@ -1,6 +1,6 @@
 import { createLinkPlugin } from './createLinkPlugin';
 import { LINK_TYPE } from './types';
-import { DEFAULTS } from './defaults';
+import { DEFAULTS, THEME as theme } from './defaults';
 
 export const pluginLink = (config = {}) => {
   return {
@@ -8,5 +8,6 @@ export const pluginLink = (config = {}) => {
     type: LINK_TYPE,
     createPlugin: createLinkPlugin,
     ModalsMap: {},
+    theme,
   };
 };