+++ packages/plugin-link/web/src/viewer.js
@@ -1,6 +1,6 @@
 import { LINK_TYPE } from './types';
 import { typeMapper } from './typeMapper';
-import { DEFAULTS } from './defaults';
+import { DEFAULTS, THEME as theme } from './defaults';
 export { typeMapper as linkTypeMapper, LINK_TYPE };
 export { default as LinkViewer } from './LinkViewer';
 
@@ -8,7 +8,7 @@ export const pluginLink = (config = {}) => {
   return {
     config: { ...DEFAULTS.config, ...config },
     type: LINK_TYPE,
-    decorator: {},
     typeMapper,
+    theme,
   };
 };