+++ packages/wrapper/web/src/EngineWrapper.jsx
@@ -38,10 +38,10 @@ class EngineWrapper extends React.Component {
   render() {
     const { strategies = [], modalSupport = true, children = {} } = this.props;
     const modifiedProps = strategies.reduce((props, strategyFunction) => {
-      const result = strategyFunction(children.props);
+      const result = strategyFunction(props);
       return { ...props, ...result };
     }, children.props);
-    const { helpers = {}, theme, locale, ModalsMap, onChange } = modifiedProps;
+    const { helpers = {}, theme, locale = 'en', ModalsMap, onChange } = modifiedProps;
     const { onRequestClose } = this.state.modalProps || {};
     if (modalSupport) {
       helpers.openModal = this.onModalOpen;