+++ CHANGELOG.md
@@ -17,6 +17,26 @@
 </details>
 <hr/>
 
+## 6.9.0-alpha.1 (Feb 25, 2020)
+6.9.0 with wix-draft
+
+## 6.9.0 (Feb 25, 2020)
+### :rocket: New Feature
+- `viewer`
+  - [#702](https://github.com/wix-incubator/rich-content/pull/702) seo mode prop for viewer render images in high resolution
+- `wrapper`
+  - [#714](https://github.com/wix-incubator/rich-content/pull/714) locale strategy added
+### :bug: Bug Fix
+- `wrapper`
+  - [#719](https://github.com/wix-incubator/rich-content/pull/719) locale strategy fixes
+- `image`
+  - [#712](https://github.com/wix-incubator/rich-content/pull/712) old image client data support #2
+- `video`
+  - [#648](https://github.com/wix-incubator/rich-content/pull/648) loader disappears before video is fully loaded
+### :house: Internal
+- `html`
+  - [#681](https://github.com/wix-incubator/rich-content/pull/681) removed htmlIframeSrc from config. Loads iframe html from bundle instead of the static file served by the consumer
+
 ## 6.8.6-alpha.3 (Feb 23, 2020)
 Add gallery fix and remove blockRendererMap